Moved Payment Method to FrameObjects
