+++ FrameSDK/src/main/java/com/framepayments/framesdk/CommonNetworkObject.kt
@@ -11,6 +11,30 @@ object FrameObjects {
         @SerializedName("line_1") val addressLine1: String? = null,
         @SerializedName("line_2") val addressLine2: String? = null
     )
+
+    data class PaymentMethod(
+        val id: String,
+        val customer: String? = null,
+        val billing: BillingAddress? = null,
+        val type: String,
+        @SerializedName("object") val methodObject: String,
+        val created: Int,
+        val updated: Int,
+        @SerializedName("livemode") val liveMode: Boolean,
+        val card: PaymentCard? = null
+    )
+
+    data class PaymentCard(
+        val brand: String,
+        @SerializedName("exp_month") val expirationMonth: String,
+        @SerializedName("exp_year") val expirationYear: String,
+        val issuer: String? = null,
+        val currency: String? = null,
+        val segment: String? = null,
+        val type: String? = null,
+        @SerializedName("last_four") val lastFourDigits: String
+    )
+
 }
 
 data class QueryItem(val name: String, val value: String?)
