+++ FrameSDK/src/main/java/com/framepayments/framesdk/customers/CustomerObjects.kt
@@ -1,7 +1,6 @@
 package com.framepayments.framesdk.customers
 
 import com.framepayments.framesdk.FrameObjects
-import com.framepayments.framesdk.paymentmethods.PaymentMethod
 import com.google.gson.annotations.SerializedName
 
 data class Customer(
@@ -16,5 +15,5 @@ data class Customer(
     @SerializedName("object") val customerObject: String?,
     @SerializedName("shipping_address") val shippingAddress: FrameObjects.BillingAddress?,
     @SerializedName("billing_address") val billingAddress: FrameObjects.BillingAddress?,
-    @SerializedName("payment_methods") val paymentMethods: List<PaymentMethod>?
+    @SerializedName("payment_methods") val paymentMethods: List<FrameObjects.PaymentMethod>?
 )
