+++ FrameSDK/src/main/java/com/framepayments/framesdk/paymentmethods/PaymentMethodResponses.kt
@@ -1,9 +1,10 @@
 package com.framepayments.framesdk.paymentmethods
 import com.framepayments.framesdk.FrameMetadata
+import com.framepayments.framesdk.FrameObjects
 
 object PaymentMethodResponses {
     data class ListPaymentMethodsResponse(
         val meta: FrameMetadata? = null,
-        val data: List<PaymentMethod>? = null
+        val data: List<FrameObjects.PaymentMethod>? = null
     )
 }
